feat: Add ability to save wordle game results
This commit adds the ability to save the results of a Wordle game to a JSON file. The results include the model used, the target word, the guesses made, whether the puzzle was solved, the number of turns, and the total time taken to play the game.

A command-line argument has been added to specify the directory where the results should be saved.